refactor: move line_sentence_chunker folder under strategies folder
