+++ tests/chunking/test_line_sentence_chunker.py
@@ -4,9 +4,9 @@ import pytest
 from textractor.entities.bbox import BoundingBox
 from textractor.entities.line import Line
 
-from ingestion_pipeline.chunking.line_sentence_chunker.chunker import LineSentenceChunker
-from ingestion_pipeline.chunking.line_sentence_chunker.config import LineSentenceChunkingConfig
 from ingestion_pipeline.chunking.schemas import DocumentMetadata
+from ingestion_pipeline.chunking.strategies.line_sentence_chunker.chunker import LineSentenceChunker
+from ingestion_pipeline.chunking.strategies.line_sentence_chunker.config import LineSentenceChunkingConfig
 
 
 @pytest.fixture
